refactor(coding-agent): flatten goal module structure
Refs: ENG-5938

fixes #2184
