+++ packages/coding-agent/src/README.md
@@ -1,6 +1,8 @@
-# Session modules
+# Source organization
 
-`AgentSession` is the public entry point and coordinates work across features. Extracted features live here, grouped by responsibility. Each owner keeps its state and transitions together and receives only the dependencies it uses.
+`src/` contains application source; tests, scripts, docs, examples, and build output stay at the package root. Put feature folders directly under `src/`, such as `goals/`, `session/`, and `kernel/`. Add another level only when a feature has distinct subparts that benefit from being grouped.
+
+`core/` currently contains most execution logic. Migrate its responsibilities into sibling feature folders as they are extracted. `AgentSession` remains the public entry point and coordinates work across features. Each feature owner keeps its state and transitions together and receives only the dependencies it uses.
 
 ## Goals
 
@@ -9,7 +11,7 @@
 | `goals/controller.ts` | Goal transitions, token and time accounting, continuation counts, and rollback checkpoints. |
 | `goals/persistence.ts` | Reading the selected branch, flushing goal records, and deciding whether a branch can receive an initial goal. |
 | `goals/commands.ts` | Parsing `/goal` arguments into typed commands. |
-| `../goals.ts` | Shared goal types, validation, serialization, and context-message formatting used by session clients. |
+| `core/goals.ts` | Shared goal types, validation, serialization, and context-message formatting used by session clients. |
 
 The controller depends on a load/save interface, an update callback, and a clock. It does not receive `AgentSession`, the agent loop, a kernel, or a UI object. Its state is read-only to callers; mutations go through named operations.
 
@@ -21,7 +23,7 @@ Three ordering rules matter during future extractions:
 - Capture completion usage, clear stale queued goal context, and then persist and publish completion. The explicit completion callback preserves this order.
 - A continuation rejected by new input restores both goal state and the accounting clock. Deferred child-work admission preserves the existing clock while restoring goal state.
 
-The public goal payload and persisted `thread_goal_state` format remain shared contracts. Internal organization does not require a new daemon command or schema.
+The shared goal types and message formatting remain in `core/goals.ts` during this extraction; their existing consumers can migrate together in a later change. The public goal payload and persisted `thread_goal_state` format remain shared contracts. Internal organization does not require a new daemon command or schema.
 
 ## Extending this structure
 
@@ -31,6 +33,6 @@ The next design review should cover input admission and turn lifecycle. The exis
 
 ## Validation
 
-Controller tests live in `test/session/goals/`. Session integration coverage remains in `test/suite/agent-session-goal.test.ts`, `test/suite/agent-session-compaction-continuation.test.ts`, and `test/goal-continuation-quiescence.test.ts`.
+Controller tests live in `test/goals/`. Session integration coverage remains in `test/suite/agent-session-goal.test.ts`, `test/suite/agent-session-compaction-continuation.test.ts`, and `test/goal-continuation-quiescence.test.ts`.
 
 Run the focused files from the coding-agent package root with the repository's prescribed Vitest command, then run `npm run check` from the repository root. Use faux providers for session tests.
